+++ src/pykeen/evaluation/rank_based_evaluator.py
@@ -8,7 +8,7 @@ import math
 import random
 import re
 from collections import defaultdict
-from dataclasses import dataclass, fields
+from dataclasses import fields
 from typing import (
     Any,
     Iterable,
@@ -37,11 +37,11 @@ from ..triples.triples_factory import CoreTriplesFactory
 from ..typing import (
     LABEL_HEAD,
     LABEL_TAIL,
-    RANK_OPTIMISTIC,
-    RANK_PESSIMISTIC,
     RANK_REALISTIC,
+    RANK_TYPE_SYNONYMS,
     RANK_TYPES,
     MappedTriples,
+    Ranks,
     RankType,
     Target,
 )
@@ -86,39 +86,6 @@ class MetricKey(NamedTuple):
         return ".".join(map(str, components))
 
 
-@dataclass
-class Ranks:
-    """Ranks for different ranking types."""
-
-    #: The optimistic rank is the rank when assuming all options with an equal score are placed
-    #: behind the current test triple.
-    #: shape: (batch_size,)
-    optimistic: torch.FloatTensor
-
-    #: The realistic rank is the average of the optimistic and pessimistic rank, and hence the expected rank
-    #: over all permutations of the elements with the same score as the currently considered option.
-    #: shape: (batch_size,)
-    realistic: torch.FloatTensor
-
-    #: The pessimistic rank is the rank when assuming all options with an equal score are placed
-    #: in front of current test triple.
-    #: shape: (batch_size,)
-    pessimistic: torch.FloatTensor
-
-    #: The number of options is the number of items considered in the ranking. It may change for
-    #: filtered evaluation
-    #: shape: (batch_size,)
-    number_of_options: torch.LongTensor
-
-    def to_type_dict(self) -> Mapping[RankType, torch.FloatTensor]:
-        """Return mapping from rank-type to rank value tensor."""
-        return {
-            RANK_OPTIMISTIC: self.optimistic,
-            RANK_REALISTIC: self.realistic,
-            RANK_PESSIMISTIC: self.pessimistic,
-        }
-
-
 def compute_rank_from_scores(
     true_score: torch.FloatTensor,
     all_scores: torch.FloatTensor,
@@ -157,13 +124,6 @@ def compute_rank_from_scores(
         number_of_options=number_of_options,
     )
 
-
-RANK_TYPE_SYNONYMS: Mapping[str, RankType] = {
-    "best": RANK_OPTIMISTIC,
-    "worst": RANK_PESSIMISTIC,
-    "avg": RANK_REALISTIC,
-    "average": RANK_REALISTIC,
-}
 
 _SIDE_PATTERN = "|".join(EXTENDED_SIDES)
 _TYPE_PATTERN = "|".join(itt.chain(RANK_TYPES, RANK_TYPE_SYNONYMS.keys()))
